[mod] creates forms folder and moves components
