+++ src/components/list-items/SegmenterArchitectureSettingsListItem/SegmenterArchitectureSettingsListItem.tsx
@@ -1,7 +1,7 @@
 import { styled } from "@mui/material";
 
 import { CollapsibleList } from "components/lists";
-import { SegmenterArchitectureSettingsGrid } from "components/styled-components";
+import { SegmenterArchitectureSettingsGrid } from "components/forms";
 
 type ArchitectureSettingsProps = {
   onModelSelect: (modelIdx: number) => void;
