+++ src/components/styled-components/index.ts
@@ -1,12 +1,7 @@
-export { AppBarOffset } from "./AppBarOffset/AppBarOffset";
-export { CustomFormSelectField } from "./CustomFormSelectField";
-export { CustomNumberTextField } from "./CustomNumberTextField";
+export { AppBarOffset } from "./AppBarOffset";
 export { ModelSummaryTable } from "./ModelSummaryTable";
-export { StyledFormControl } from "./StyledFormControl";
 export { TrainingHistoryPlot } from "./TrainingHistoryPlot";
 export { ImageSortSelection } from "./ImageSortSelection";
-export { SegmenterArchitectureSettingsGrid } from "./SegmenterArchitectureSettingsGrid";
-export { ClassifierArchitectureSettingsGrid } from "./ClassifierArchitectureSettingsGrid";
 export { Cursor } from "./Cursor";
 export { KeyboardKey } from "./KeyboardKey";
 export { LogoLoader } from "./LogoLoader";
